Sanitize field names from Mappings is JournaldLogHandler
If the caller is using field names that journald can't accept for any
reason, transform those into something journald can accept, and then log
it that way. Fixes #35.

With this change, there is a possibility that multiple keys will be
sanitized to the same journald-safe string, and only some of them will
be logged. But this should be rare, and if it does happen, probably the
user is using this calling convention for log message keyword
formatting, and isn't so concerned with having their data logged as
journal fields. So I think keeping the implementation as simple as
possible is more valuable than worrying about that corner case.